% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_programming_knapsack.R
\name{dynamic_programming_knapsack}
\alias{dynamic_programming_knapsack}
\title{dynamic_programming_knapsack}
\usage{
dynamic_programming_knapsack(x, W)
}
\arguments{
\item{x}{data frame.}

\item{W}{A number.}
}
\value{
result containing value and elements in \code{list}.
}
\description{
This function solve knapsack using dynamic programming approach.
}
