% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{data frame.}

\item{W}{A number.}

\item{fase}{A bool}
}
\value{
result containing value and elements in \code{list}.
}
\description{
This function solve knapsack using greedy approach.
}
