% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data frame.}

\item{W}{A number.}
}
\value{
result containing value and elements in \code{list}.
}
\description{
This function solve knapsack using brute force approach.
}
