% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic solution for knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with 2 columns which include w (weight)and v (value)}

\item{W}{Represent the total capacity of the knapsack}
}
\value{
A list which includes the total value of all the items and all the elements that are in the knapsack.
}
\description{
This algorithm will does not give an exact result, but it will reduce the computational complexity considerably.
}
